1613a - Changes for moving and updating filtering rule to handle categorical data
