+++ projects/_03_independent_cqc/unittest_data/polars_ind_cqc_test_file_schemas.py
@@ -801,27 +801,6 @@ class ArchiveFilledPostsEstimates:
 
 @dataclass
 class CleanFilteringUtilsSchemas:
-    add_filtering_column_schema = pl.Schema(
-        [
-            (IndCQC.location_id, pl.String()),
-            (IndCQC.ascwds_filled_posts_dedup_clean, pl.Float64()),
-        ]
-    )
-    expected_add_filtering_column_schema = pl.Schema(
-        list(add_filtering_column_schema.items())
-        + [
-            (IndCQC.ascwds_filtering_rule, pl.String()),
-        ]
-    )
-    update_filtering_rule_schema = pl.Schema(
-        [
-            (IndCQC.location_id, pl.String()),
-            (IndCQC.ascwds_filled_posts_dedup, pl.Float64()),
-            (IndCQC.ascwds_filled_posts_dedup_clean, pl.Float64()),
-            (IndCQC.ascwds_filtering_rule, pl.String()),
-        ]
-    )
-
     aggregate_values_to_provider_level_schema = pl.Schema(
         [
             (IndCQC.location_id, pl.String()),
