+++ utils/column_values/categorical_columns_by_dataset.py
@@ -25,6 +25,8 @@ from utils.column_values.categorical_column_values import (
     Dormancy,
     EstimateFilledPostsSource,
     InAscwds,
+    JobGroupLabels,
+    JobRoleFilteringRule,
     MainJobRoleID,
     MainJobRoleLabels,
     PrimaryServiceType,
@@ -211,6 +213,14 @@ class EstimatedIndCQCFilledPostsCategoricalValues:
     estimate_filled_posts_source_column_values = EstimateFilledPostsSource(
         IndCQC.estimate_filled_posts_source
     )
+
+
+@dataclass
+class EstimatedIndCQCFilledPostsByJobRoleCategoricalValues:
+    job_role_filtering_rule_column_values = JobRoleFilteringRule(
+        IndCQC.job_role_filtering_rule
+    )
+    job_group_column_values = JobGroupLabels(IndCQC.main_job_group_labelled)
 
 
 @dataclass
